Updated EntryCount and DumpDistribution to reflect only hash entries >= BeginAddress.
