+++ cs/src/core/Index/FASTER/FASTERBase.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT license.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
@@ -692,93 +690,5 @@ namespace FASTER.core
 
             return (found == expected);
         }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        protected virtual long GetEntryCount()
-        {
-            var version = resizeInfo.version;
-            var table_size_ = state[version].size;
-            var ptable_ = state[version].tableAligned;
-            long total_entry_count = 0;
-
-            for (long bucket = 0; bucket < table_size_; ++bucket)
-            {
-                HashBucket b = *(ptable_ + bucket);
-                while (true)
-                {
-                    for (int bucket_entry = 0; bucket_entry < Constants.kOverflowBucketIndex; ++bucket_entry)
-                        if (0 != b.bucket_entries[bucket_entry])
-                            ++total_entry_count;
-                    if (b.bucket_entries[Constants.kOverflowBucketIndex] == 0) break;
-                    b = *((HashBucket*)overflowBucketsAllocator.GetPhysicalAddress((b.bucket_entries[Constants.kOverflowBucketIndex])));
-                }
-            }
-            return total_entry_count;
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="version"></param>
-        protected virtual string DumpDistributionInternal(int version)
-        {
-            var table_size_ = state[version].size;
-            var ptable_ = state[version].tableAligned;
-            long total_record_count = 0;
-            Dictionary<int, long> histogram = new Dictionary<int, long>();
-
-            for (long bucket = 0; bucket < table_size_; ++bucket)
-            {
-                List<int> tags = new List<int>();
-                int cnt = 0;
-                HashBucket b = *(ptable_ + bucket);
-                while (true)
-                {
-                    for (int bucket_entry = 0; bucket_entry < Constants.kOverflowBucketIndex; ++bucket_entry)
-                    {
-                        if (0 != b.bucket_entries[bucket_entry])
-                        {
-                            var x = default(HashBucketEntry);
-                            x.word = b.bucket_entries[bucket_entry];
-                            if (tags.Contains(x.Tag) && !x.Tentative)
-                                throw new FasterException("Duplicate tag found in index");
-                            tags.Add(x.Tag);
-                            ++cnt;
-                            ++total_record_count;
-                        }
-                    }
-                    if (b.bucket_entries[Constants.kOverflowBucketIndex] == 0) break;
-                    b = *((HashBucket*)overflowBucketsAllocator.GetPhysicalAddress((b.bucket_entries[Constants.kOverflowBucketIndex])));
-                }
-
-                if (!histogram.ContainsKey(cnt)) histogram[cnt] = 0;
-                histogram[cnt]++;
-            }
-
-            var distribution =
-                $"Number of hash buckets: {{{table_size_}}}\n" +
-                $"Total distinct hash-table entry count: {{{total_record_count}}}\n" +
-                $"Average #entries per hash bucket: {{{total_record_count / (double)table_size_:0.00}}}\n" +
-                $"Histogram of #entries per bucket:\n";
-            foreach (var kvp in histogram.OrderBy(e => e.Key))
-            {
-                distribution += $"  {kvp.Key} : {kvp.Value}\n";
-            }
-
-            return distribution;
-        }
-
-        /// <summary>
-        /// Dumps the distribution of each non-empty bucket in the hash table.
-        /// </summary>
-        public string DumpDistribution()
-        {
-            return DumpDistributionInternal(resizeInfo.version);
-        }
-
     }
-
 }
